requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0